Random Quote Generator
This project contains multiple functions that is using an array of objects to return a random selected quote to the html page. The objects are having multiple properties in them, and if statements are in charge of controlling which object properties to print. The html background color changes with changing to new quotes, or at intervals of 10 seconds.